+++ packages/client/src/features/FooterLiderBoard/index.tsx
@@ -1,0 +1,17 @@
+import { useNavigate } from 'react-router-dom';
+import { Button } from '@mui/material';
+
+export default function FooterLiderBoard() {
+  const navigator = useNavigate();
+
+  return (
+    <Button
+      variant="default"
+      onClick={() => {
+        navigator(-1);
+      }}
+    >
+      Назад
+    </Button>
+  );
+}
