+++ packages/client/src/features/HeaderLiderBoard/index.tsx
@@ -1,0 +1,7 @@
+export default function HeaderLiderBoard() {
+  return (
+    <div>
+      <h1>ТОП 10</h1>
+    </div>
+  );
+}
