+++ packages/client/src/features/LiderItem/styles.module.scss
