+++ packages/client/src/widgets/LiderBoardPage/style.module.scss
